[SYCL][CUDA][HIP] Deprecate context interop for CUDA and HIP (#10975)
The `sycl::context` does not map clearly to a native context for CUDA
and HIP backends. This is especially true now that we are adding support
for multi device context https://github.com/intel/llvm/pull/10737 . It
would be good to start this deprecation process. PRs to oneMKL and
oneDNN to follow