[SYCL][Graph] Verbose E2E error messages
The current error output from Graph E2E tests isn't
particularly informative on the data that caused an
assert to fail verification. As asserts are done on the
whole container, rather than individual values. This
change verifies each value individually so that a
more informative error message can be printed when
the assert fails. This is helpful for debugging
issues that appear in CI but are hard to reproduce locally.

Co-authored-by: Maxime France-Pillois <maxime.francepillois@codeplay.com>