[SYCL] Provide initialization for variable
As far as the compiler is concerned, the variable `x` is uninitialized
on some control-flow paths and is passed to the function
`(ex|in)clusive_scan_over_group`.

In terms of LLVM, the function parameter that `x` is passed to is likely
marked `noundef`. This can lead the compiler to take advantage of
`undef` rules and assume that the control path in which `x` is
defined must happen (otherwise it'd be undefined), and it can
essentially remove the guarding `i < N` around the memory accesses. This
is, of course, dangerous.

Though, in practice, this is often avoided because the scan function is
inlined before LLVM makes that assumption, this should not be relied
upon. I've chosen to zero-initialize it. The value of `init` cannot be
used because it may be out of bounds for the type pointed to by `InPtr`,
which can introduce new UB into the program.