[Driver][SYCL] Enable early AOT abilities when creating objects (#11130)
Adds support for enabling the ability to perform a full device
compilation and link. Use the -fno-sycl-rdc option when also compiling
to object with -c. This will trigger the additional device linking steps
to be performed after the device compilation is completed.

Currently only supported for AOT enabled targets for spir64.

Upon consumption of these new fat objects, the driver will scan for
these unique binaries and instead of going through the device link,
these binaries will be sent directly to the host link step to be added
to the final executable. This is done by introducing a few triple
architecture values to designate target images in the fat objects
(spir64_gen_image, spir64_fpga_image, spir64_x86_64_image)

Performs some refactoring of the device link code, allowing for a common
platform for compile and link to access.