fpga_mem/fpga_datapath clarifications (#11387)
Changes:
* Added default constructor to `fpga_mem` and `fpga_datapath` in order
to avoid initialing the underlying data in the objects when they are
instantiated in function scope. This allows the compiler to optimize in
more cases as it avoids creating am additional write to the object. More
over it aligns the initialization behavior of a a type being
instantiated at function scope without `fpga_mem`/`fpga_datapath` and
with. They are now the same.
* Added tests to check for proper initialization behavior.
* Documented default constructors in SYCL specs
* Documented copy and move constructors and assignment operators
* Updated the nesting behavior of `fpga_mem` and `fpga_datapath` to
align it with what is currently supported through annotations.

---------

Co-authored-by: Steffen Larsen <steffen.larsen@intel.com>
Co-authored-by: Joe Garvey <joseph.garvey@intel.com>