[SYCL] Fix SYCL kernel lambda argument type detection
We have a helper which is used to extract a type of the first SYCL
kernel lambda argument to do some error-checking and special handling
based on that.

That check, however, was missing a case when a kernel lambda is also
accepting `kernel_handler` argument, always falling back to a suggested
type in that case. This led to a situations where we couldn't compile
code like:

```c++
sycl::queue q;
q.parallel_for(sycl::range{1}, [=](sycl::item<1, false>, kernel_handler) {});
```

This patch adds extra specializations of some internal helpers to fix
the error.

This is a follow-up from intel/llvm#11625