[Driver][SYCL] Adjust debug format for SYCL device on Windows
Setting debug information is determined by using -g, as opposed to
earlier representation using -Zi specifically for Windows.  The use of
codeview is dependent on the target.

When we know that the host target is MSVC based, use the host toolchain
setting of the debug format as opposed to the default toolchain which
defaults to dwarf.