[SYCL][Graph][L0] Reset L0 events for UR sync-points (#11553)
The [Unified Runtime
sync-points](https://oneapi-src.github.io/unified-runtime/core/EXP-COMMAND-BUFFER.html#sync-points)
used to represent SYCL-Graph edges have a bug in the implementation of
the Level Zero adapter. The L0 events used to implement the sync-points
are not reset to the non-signaled state after an enqueue of the graph.

Fixed in the UR L0 adapter change
https://github.com/oneapi-src/unified-runtime/pull/965 which resets the
UR events to the non-signaled state at the end of the L0 command-list.