[CUDA][HIP] Move adapter specific testing out of CTS
This patch moves the `test-adapter-cuda` and `test-adapter-hip` test
suites out of the `test/conformance` directory. The goal of this move is
to ensure that all adapters will be tested with the same set and number
of conformance tests, making conformance pass rates directly comparable
between adapters. Up to now the CUDA and HIP adapters have included
additional unit tests of internals which skews pass rates.