[CTS] Add UR_SYCL_LIBRARY_DIR CMake variable
The CTS is already dependant on the DPC++ compiler to generate programs
inputs for the program, kernel, and enqueue test suites specified via
the `UR_DPCXX` CMake variable. If the DPC++ compiler is not installed on
the system the executables it outputs will likely fail to find the SYCL
runtime library when executed, breaking the generation of CTS program
inputs. The patch introduces the `UR_SYCL_LIBRARY_DIR` CMake variable
enabling the user to specify the path to the SYCL runtime library for
use when generating CTS program inputs.