[SYCL][Driver] Fix -fgpu-rdc option for CUDA (#11420)
This commit fixes an issue with -fgpu-rdc for NVPTX targets in the SYCL
driver. When the driver calls the `ptxas` assembling job in relocatable
device code compilation mode, we pass `-c` to the command-line to get a
relocatable code ELF instead of a statically fully linked object. This
is the correct behaviour for the purpose of CUDA separate compilation
flow with RDC, but in SYCL it does not make sense for us to do so
because we do RDC at a bitcode level rather than PTX. We link the IR
modules with llvm-link way earlier and the linking method based on
`-fgpu[no]-rdc` is determined there. Hence, we always want a fully
linked object from `ptxas` in SYCL. This simplifies the need for adding
a `nvlink` job in the pipeline.

Related github issue: https://github.com/intel/llvm/issues/8690