[ESIMD] Fix the declarations and uses of min and max functions
min and max() functions are absoletely legal to use in normal
conditions,
e.g. std::min(a, b), std::max(a, b), or esimd::max(simd1, simd2).
That may become a problem if 'windows.h' or 'minwindef.h' is included
and min and max MACRO are defined.

This patch fixes potential problems with that in ESIMD and std::simd
headers.
Also, adds 1 E2E test and 1 compile-only test for min, max.

Signed-off-by: Klochkov, Vyacheslav N <vyacheslav.n.klochkov@intel.com>