[SYCL][Doc] Add kernel compiler extension spec (#11331)
Add a proposed extension specification for the "kernel compiler", a
replacement for the existing "online_compiler" extension. There are two
related extensions specifications:

* sycl_ext_oneapi_kernel_compiler: Defines the API framework for online
compilation of kernels from a source string. This framework allows for
kernels to be written in different possible languages.

* sycl_ext_oneapi_kernel_compiler_opencl: Defines the support for online
compiled kernels that are written in OpenCL C.

---------

Co-authored-by: John Pennycook <john.pennycook@intel.com>