[SYCL] Address unused-private-field error  (#11872)
During post-commit testing, the self build is failing with an
`unused-private-field` error naming the `kernel_bundle_impl.Language`
member as being unused. That member is only used in asserts right now in
the existing code, so I am assuming this is the problem. Here am I
adding an explicit check that should avoid the error.