[SYCL] Change return type of abs_diff
As of https://github.com/KhronosGroup/SYCL-Docs/pull/458 the return type
of abs_diff has been changed to be the same as the input arguments,
which in turn corresponds to how the extended OpenCL instruction set
defines `s_abs_diff`. This commit changes the preview builtins to use
the correct type as well as define `s_abs_diff` in accordance with
the OpenCL extended instruction set.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>