Change comment for alloca insertion
Before the global-offset optimization, non-offset
and offset kernels called the same function with
an added offset argument.
The non-offset versions allocated an array with
3 zeros in addrspace(5) and passed it as an
argument to the function calls containing
calls to the intrinsic.
Since the allocation and filling is not possible
in addrspace(4), the offset version adds an extra
alloca in addrspace(5) and does an addrspace cast
to addrspace(4).
This is not required anymore since the non-offset
version does not perform an alloca of zeros
anymore. However, the optimization is rendered
almost useless as the global offset feature is
deprecated.
Adapt the comment to reflect this.