[SYCL] Change return type of abs_diff
As of KhronosGroup/SYCL-Docs#458 the return type of abs_diff has been
changed to be the same as the input arguments, which in turn corresponds
to how the extended OpenCL instruction set defines s_abs_diff. This
commit changes the preview builtins to use the correct type similar to
how `sycl::abs` does.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>