[SYCL] Add documentation and change default val for min range x to enable range rounding (#11823)
Some users found that 1024 was too conservative a min X dim to enable
range rounding. This lowers that val.

It also adds some documentation for SYCL range rounding, which had been
missing. Perhaps this should be in its own file. Let me know where to
put it and I will do so.