[ESIMD] Add non-experimental version of fence() for DG2 and PVC
The new esimd::fence() accepts 3 template enum parameters and is
usable only on DG2 and PVC.
It implements the subset of functionality supported by experimental::esimd::lsc_fence().
The functionality that was not included into the public version is
a) not supported by hardware and/or GPU drivers
b) too specific and is too risky to use now.

Signed-off-by: Klochkov, Vyacheslav N <vyacheslav.n.klochkov@intel.com>