[SYCL] Add environment variable to disable in-memory program caching (#11751)
This PR adds an environment variable `SYCL_CACHE_IN_MEM` to control the
in-memory caching of programs. Currently, every program/kernel is saved
in the global `KernelProgramCache`, which means that every
program/kernel will not be released until end of the program when the
destructor of `KernelProgramCache` is ran. By enabling this environment
variables, caching is not performed and the resources program and
kernels use are freed after use.