[SYCL] Fix integer vec conversions (#11821)
This is a bugfix for regressions caused by intel/llvm#11770

The following cases are fixed/updated:
- Signed to unsigned (and vice versa) conversions are made to always go
through per-element path, because native SPIR-V instruction behavies
differently from what SYCL CTS expects. Strictly speaking, this could be
a CTS issue, because `vec::convert` is underspecified (see
KhronosGroup/SYCL-Docs#492), but until the spec is clarified we will
stick with existing behavior;
- Conversions to bool are made to always go through per-element path,
because under the hood bool is represented as an integer and there is no
guarantee that regular integer conversion will produce the right bit
patterns for `true` and `false` which are not defined by any
specification;

Added an on-device test to check that `vec::convert` for integer types
behaves correctly and consistently on both host and device.