[Driver][SYCL][FPGA] Improve AOCX archive behaviors with -fsycl-link=image (#11592)
Use of -fsycl-link=image allows for FPGA users to create a full device
binary and store it in an archive to be used later. This archive can
then be picked up and linked in. The information stored in the archive
was not sufficient, causing the device binary to not be 'visible' when
linked in.

Improve the situation here by saving all of the information provided by
the wrapping step (including symbols) in the archive, but also be sure
that this entire final binary is saved as it would be if we were
performing a full E2E compilation.

To accomodate the ability to pass around the needed symbols and
properties that are used when wrapping the final binaries for any AOCR
or AOCX type archives, update the way this information is passed around.
Multiple binaries are added to the archive, including the full wrapped
.bc file that is gathered and passed along to any subsequent wrapping
sequence that occurs.

---------

Signed-off-by: Lu, John <john.lu@intel.com>
Co-authored-by: Lu, John <john.lu@intel.com>