[NFC][Driver] Remove inadvertent append for the host link (#11922)
When performing the task involving early-AOT behaviors, an additional
virtual function override of appendLinkHostActions was added, which was
returning the end of the action list. This did not have an overall
impact on the SYCL compilation flow, so it was unnoticed. When combining
with the OpenMP interop offloading steps, this was causing an additional
invalid object/device to be added to the host link.

This can only be tested with the old offloading model, which only exists
with the SYCL toolchain on the intel/llvm branch.