[SYCL] fix a behavior mismatch in ann_ptr operator (#11904)
Fixed two issues:
1. For compound operators, there is an error in the specs. The `tmp =
tmp Op rhs` statement is wrong since Op is defined as `+=` and for `+=`
it will be `tmp = tmp += rhs`. Correct the format
2. For compound operators and inc/dec operators, the old implementation
is using the non-compound operator which is wrong.
i.e. If T is a customer type with only `+=` operator being defined, when
applying `+=` on `annotated_ref<T>`, the previous implementation errors
out since there is no `+` defined in T.
3. Added the missing `const` to inc/dec operators