[SYCL][Bindless][Exp] Allow `read_image/mipmap` to return user types (#11961)
- `read_image` and `read_mipmap` now allow user-defined types to be
returned.

- The `write_image` test case for user-defined types is updated to
include both reading and writing of user-defined types.

- An additional test is added for reading user-defined types where the
image type is a mipmap.

 - The extension document has been updated to reflect the above changes.

- The `write_image` function now checks whether a user-defined type was
passed to it, and converts the input data accordingly. If the type
passed has a specific NVPTX backend intrinsic, it will use that,
otherwise it will write the user data type bit-cast to a variant of
unsigned integer.

- This patch marks the bump to revision `5.0` of the bindless images
specification document.