[SYCL][E2E] Fix preview flag testing for non-HIP and non-CUDA devices (#11978)
The preview flag testing for E2E testing is currently broken.
https://github.com/intel/llvm/pull/11929 attempts to fix it but has
shown that some breakages in CUDA and HIP have been hidden behind the
broken testing. To allow for proper testing while these failures are
being investigated, this PR implements the fixes from
https://github.com/intel/llvm/pull/11929 with the caveat that they do
not apply when HIP and CUDA devices are part of the testing. This is a
temporary solution and does *not* replace
https://github.com/intel/llvm/pull/11929.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>
Co-authored-by: Chris Perkins <chris.perkins@intel.com>