[opencl-aot] Fix regression in error reporting (#12022)
This patch fixes a regression in error reporting, previously compiler
build log was printed to `std::cerr`, but now it doesn't print by
default. It makes it difficult for developers to understand what's going
on. So now the compiler build log is printed to `std::cerr` if there is
an error message, and to verbose logs if compilation was successful but
there is something in the compiler build logs.