Fixing Intrinsic::ptr_annotation (#2235)
There is a fundamental misalignment between when translating llvm.ptr.annotation to SPIRV's OpMemberDecorate.

llvm.ptr.annotation is applied onto a pointer.

OpMemberDecorate is applied on a type, which means that all variables of this type will have this Decoration.

This change fixes this misalignment by translating llvm.ptr.annotation to SPIRV's OpDecorate instead, which instead places the decoration onto the pointer.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/3770469