[SYCL] Fix issue of acquring kernel twice (#11953)
In #11751, ref counting of kernels objects was changed to be more
accurate in order to allow for in-memory caching to be disabled. When
getting a kernel form the cache, the ref count the kernel handle is now
incremented (when caching is enabled). Thus, a method like
`ProgramManager::getOrCreateKernel` will increment the ref count of the
kernel it gets. However, in `enqueueImpKernel`, when enqueuing a kernel
with a kernel bundle,`ProgramManager::getOrCreateKernel` is called
twice, first indirectly by:


https://github.com/intel/llvm/blob/c43a90f28eebfcdf1bc1d55430485e2834790a60/sycl/source/detail/scheduler/commands.cpp#L2527-L2528

and second directly by:


https://github.com/intel/llvm/blob/c43a90f28eebfcdf1bc1d55430485e2834790a60/sycl/source/detail/scheduler/commands.cpp#L2538-L2548

This means that the ref count of the acquired kernel is incremented
twice, yet the rest of the function will only free once, which leads to
a leak of the kernel. As the second comment and asserts say, the only
need for the second call to `getOrCreateKernel` is to fetch the mutex
associated to the cached kernel retrieved from the first call, so this
PR adjusts `get_kernel` to save this mutex and forgo this extra
`getOrCreateKernel` call and unintentional additional ref count.