[SYCL] Optimize int-convert.cpp test
The test sporadically timed out on Windows because default implicit
context is disabled in there.

Due to the fact that separate `queue` object was create for every
kernel, it caused full re-JIT of the whole device code.

Refactored the test to use a single queue object explicitly to have
context and all associated caches to be shared between all kernels in
the test.

This resolves intel/llvm#12011