Can apply MemoryINTEL annotation multiple times on same variable (#2230)
According to SPIRV spec only one instance of a specific decoration can be on a SPIRV-Id, unless the decoration explicitly allows it.

LLVM IR has no such limitation.

MemoryINTEL can have two annotations on the same pointer in LLVM IR. When it comes to translating this to SPIRV, we can filter out the less precise annotation (DEFAULT) and only keep one that will be translated to a SPIRV decoration.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/ff48393