[Driver][SYCL] Do not pass coverage and profiling opts to device compile
When building with profiling or code coverage options, do not enable
these passes for the device compilation.  When used, and additional
diagnostic will be emitted stating how the options is not being used for
the given target.

Example: clang++ -fprofile-instr-generate -fsycl file.cpp
clang++: warning: ignoring '-fprofile-instr-generate' option as it is not currently supported for target 'spir64-unknown-unknown' [-Woption-ignored]