[SYCL][ESIMD] Implement unified memory API - atomic_update(acc,...) with two operands (#12055)
This implements atomic_update with two operands and accessors. 

We only need four overloads to cover all cases because we can rely on
the simd_view to simd implicit conversion operator. The behavior of
every possible combination is locked down in a compile time test, and I
manually verified the all the combinations are correct and nothing is
missed.

I also manually tested on PVC.

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>