[SYCL][ESIMD][E2E] Fix Arc postcommit (#12115)
The tests will fail if they actually run, but the driver used in CI is
old enough to hit the runtime driver check so they return as passed
before actually doing anything. Just mark them as unsupported for now to
simplify things. We already have an internal tracker for the root issue.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>