[SYCL] Evict program cache on PI_ERROR_OUT_OF_RESOURCES (#11987)
This PR adds resource cleanup mechanisms when
`PI_ERROR_OUT_OF_RESOURCES` occurs on `piProgramBuild` and
`piProgramLink`. Currently. the whole cache is cleared, but this
behavior can be extended in the future. In order to ensure thread safety
of the cache clearing mechanism, the maps of the cache now hold their
`BuildResult`'s in `shared_ptr`'s, which allows the cache to be cleared
even if there are still threads that still hold references to a
`BuildResult`.