[SYCL][Fusion] Pass accessor element size in bytes to internalization pass (#12108)
This PR removes some remnants of the opaque pointer transition from the
internalization pass. For private internalization, we insert an `alloca`
instruction to model a work-item's private storage. Currently, we have
to infer the accessor's element type by looking at its users. Now, we
simply pass the element size as metadata to the pass, and insert an
`i8`-`alloca` with the desired size in bytes. I'm using separate
`LocalSize` (= number of buffer elements associated with each work-item
resp. work-group) and `ElemSize` here for now as the remapping of GEP
instructions still works in terms of number-of-elements. If in the
future these are rewritten to perform `i8*` arithmetic as well, we could
go back to only passing a single size (then in bytes instead of number
of elements).

---------

Signed-off-by: Julian Oppermann <julian.oppermann@codeplay.com>