[Driver][SYCL] Update aocx/aocr binary usage
Updates were performed for AOCX/AOCR usage and creation to better
support the symbols and properties retention when using -fsycl-link.
These changes were not fully Windows supported due to the restrictions
in creating libraries.

Fix this situation to create objects that can be added to the Windows
library as opposed to just a bundled IR file.