[SYCL][ESIMD][E2E] Add DG2 atomic_update tests (#12131)
Add atomic_update tests for DG2. I manually tested these on PVC Linux
and DG2 Linux/Windows.

For the normal accessor and USM tests, all tests work on DG2 with no
changes, so I just moved the PVC tests to be DG2/PVC tests instead of
making new files since there's no code change required besides the
`REQUIRED` line.

For the SLM and SLM accessor tests, we have to skip 64-bit types as
those are not supported for LSC SLM atomics on DG2, if we try to use
them we get an error explicitly saying that. So for these I did
something similar to what I did for block_load/block_store and just have
an if check to skip that case.

The diff is a little bit weird, some of the file matching is a bit
wrong, so you might have to look for the correct files to compare in the
file list.

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>