Return device total global memory for MaxAllocSize
The SYCL specifiction inherits an OpenCL query for the maximum single
allocaton that can be made, and requires that at least one-quarter of
the device's reported global memory size can be allocated.

CUDA doesn't really have such a limitation as far as I can tell, and
will happily allocate anything up to the total size of global memory
in one go. This means returning that size is a reasonable guess for
MaxAllocSize.