[sycl-post-link][Driver] Pass in -emit-only-kernels-as-entry-point option by default (#12060)
sycl-post-link is one of the tools invoked during compilation of SYCL device code. One of the options passed to this tool is ‘emit-only-kernels-as-entry-points’. This option was introduced to stop treating `SYCL_EXTERNAL` functions as entry points. However, it was turned on by default only for FPGAs and turned off for other backends. We are noticing that there are no tangible benefits to treating `SYCL_EXTERNAL` functions as entry points and that treating them as entry points results in unnecessary device code bloating. This is a pain point reported by several customers. We propose to turn on `
emit-only-kernels-as-entry-points` by default for all Intel backends. However, we want to expose this option to users and allow them to turn it off if required. Hence we propose the following command line option
to be added into compiler:  -f[no]-sycl-remove-unused-external-funcs // Linux and Windows
This option takes no arguments

Thanks
---------

Signed-off-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>