[SYCL][Fusion] Mark unsupported on accelerator
Kernel fusion is currently not supported for SYCL devices of device type
accelerator, as devices such as FPGAs do not support JIT compilation at
runtime.

Mark all end-to-end tests related to kernel fusion as unsupported for
devices of type accelerator.

Also use the newly created file to make the fact that kernel fusion is
currently unsupported on Windows more explicit.

Signed-off-by: Lukas Sommer <lukas.sommer@codeplay.com>