[SYCL][Doc] Rename scope-specific variables (#11166)
The semantics of work-group-local variables are aligned with C++
thread-local variables but use a different suffix. This has led to
confusion, as some readers assume the "local" in work-group-local refers
to the local address space

After considering many alternatives, we settled on the suffix "specific"
to describe this concepts: work-group-specific variables are associated
with a specific work-group. In future, we expect device-global variables
to be renamed to device-specific variables for consistency.

---------

Signed-off-by: John Pennycook <john.pennycook@intel.com>