Resolve issues about decomposed and optimized args
We decided that free function kernels can accept all legal argument
types, even those types that need to be decomposed.  Remove the open
issue and add an implementation note.

Do the same for the issue about arguments that have been optimized
away.  It seems like this can be implemented with similar logic that
will be needed for the decomposed arguments.  However, this
implementation note does not require implementations to perform these
optimizations.  It merely points out the ramifications.