Classify languages as text vs. byte
It's not clear what it would mean to create a kernel bundle from raw
bytes if the kernel was written in a textual language (e.g. OpenCL).
It seems like we would need to define the character encoding, and that
seems like a bad idea.

Instead, classify each source language as either text format or binary
format.  We now require the application to use the overload of
`create_kernel_bundle_from_source` that matches the format.