[hip] Fix HSA headers lookup
ROCm installations prior to v6 don't respect the traditional
installation layout and install the HSA headers to
`$PREFIX/hsa/include/hsa` whereas in rocm6 it looks like they're putting
it in the right place at `$PREFIX/include/hsa`.

That cleanup from AMD is good news, but it means that our workarounds
now break and we need to check both places.