[vecz] Merge and fix call instantiation LIT tests
Some of the 'failure' tests were broken for reasons seen in previous
commits: the test was actually passing, but the incorrect use of
FileCheck was satisfying the XFAIL.

These tests all use the same input, so are now merged. Those which test
cases we can packetize, test that codegen. Those that can't, check we
don't produce a vectorized function.