[SYCL][Fusion][NoSTL] Replace YAML I/O for SYCLModuleInfo by metadata-backed approach (#12262)
For testing purposes, the `SYCLModuleInfo` data structure (and its
dependents) implemented LLVM's YAML traits for easy (de-)serialization.
However, as the YAML infrastructure requires the participating structs
to support dynamic construction of their instances (e.g. growing a
vector of argument types during parsing), this is detrimental to our
goal of dropping STL classes from the fusion interface.

This PR rewrites the affected lit tests by encoding the information to
be injected as a named metadata in the IR module. Only tests for the
kernel fusion pass actually require the module info; it could be dropped
without replacement from the internalization and CP tests. We also no
longer encode the binary format information as the tests were not
sensitive to it. The named metadata is not intended to persist across
the pipeline, and is currently erased right after parsing. The YAML
export was replaced by a simpler (but still YAML-inspired) textual dump
of the module info.

*This PR is part of a series of changes to remove uses of STL classes in
the kernel fusion interface to prevent ABI issues in the future.*

---------

Signed-off-by: Julian Oppermann <julian.oppermann@codeplay.com>