[SYCL] Limit the directories that are searched when loading plugins
Currently the default search order is used when loading dependencies of the
plugins (these depencies include the Level Zero loader and the ICD
loader) and that list includes current directory and some other
directories which are not considered safe. This patch limits the list of
directories when loading the dependencies of the plugins.
See: https://learn.microsoft.com/en-us/windows/win32/api/libloaderapi/nf-libloaderapi-loadlibraryexa
for reference.