[SYCL][Matrix tests]make multi_ptr access direct inside the kernel (#12351)
This change explicitly defines multi_ptr inside the kernel instead of
outside the kernel. Before, we rely on the runtime to correctly treat
the lambda capture clause and create correct private copies for each of
these variables. With this change, we are making sure that these
variables are correctly captured (private copies) inside the kernel.