[SYCL] Limit the directories that are searched when loading dependencies of the plugins (#12336)
Currently the default search order is used when loading dependencies of
the plugins (these dependencies include the Level Zero loader and the
ICD loader for opencl and level zero plugins respectively) and that list
includes current directory and some other directories which are not
considered safe. This patch limits the list of directories when loading
the dependencies of the plugins. See:
https://learn.microsoft.com/en-us/windows/win32/api/libloaderapi/nf-libloaderapi-loadlibraryexa
for reference.

---------

Co-authored-by: aelovikov-intel <andrei.elovikov@intel.com>