[SYCL][Graph] Fix post-commit CI after PR 11472
Ignore unused `Size` parameter from
`memory_manager::ext_oneapi_fill_cmd_buffer`.

This was introduced in PR https://github.com/intel/llvm/pull/11472
and passed as `Req->MMemoryRange` from the command-group requirement
but not used for calling the underlying UR function, leading to a
`-Wunused-parameter` error.

It may be that this is needed to support buffers of dimension greater
than 1, which there are currently no graphs tests for. This will
be investigated as a follow-up once post-commit CI is unblocked
with this PR.