[L0] move platform cache into the adapter structure
The platform cache is a global variable used exclusively by the
L0 adapter, and it's protected by a loosely-associated spin lock.
However, its destruction is associated with the lifetime of the
adapter structure and is deleted the first time adapter refcount
reaches 0. This was causing issues whenever the adapter was
initialized and destroyed multiple time inside of a single process,
which, for example, happens during tests.

This patch fixes the above problem by moving the platform cache
from the global state into the adapter structure. This allows
for a simpler implementation that no longer requires an explicit
lock and instead uses lazy loading (std::call_once).

With this patch, all platform tests are now passing for L0.
Closes #824