[SYCL][Fusion][NoSTL] Encode argument descriptor using fixed-size arrays (#12355)
Introduces the `jit_compiler::DynArray` template to represent a
fixed-sized-but-dynamically allocated array with an
`std::array`-compatible interface. `DynArray` is then used to replace
the uses of `std::vector` in the `SYCLArgumentDescriptor` struct. Client
code was updated to either determine the number of arguments beforehand,
or use dynamically-growable containers to collect the relevant info
locally and then copy it over to a `DynArray` instance.

*This PR is part of a series of changes to remove uses of STL classes in
the kernel fusion interface to prevent ABI issues in the future.*

Signed-off-by: Julian Oppermann <julian.oppermann@codeplay.com>