[HIP] Changes for device_globals and enable tests (#12165)
Enable `device_global` for AMDGPU. Related to
https://github.com/oneapi-src/unified-runtime/pull/1186 .

In order for `hipModuleGetGlobal` to see a global symbol we need to not
make the visibility of the symbol hidden. Perhaps changing this as a
default is not always a good idea. Could also do this in the SYCL
headers using an attribute to specify the visibility of the
`device_global` var